python-server/tf/requirements.txt
grpcio
grpcio-tools
tensorflow
tensorflow-serving-api